feat: Issue #590 - Phase 0-2: エラーハンドリング系ファイル統合完了
- context_analyzer_core.py と context_analyzer_reports.py を統合
- 3ファイル → 1ファイル（67%削減）で開発効率向上
- ティア別ファイルサイズ制限（Important Tier: 500行）準拠
- 後方互換性確保（廃止予定クラス名でのアクセス維持）

統合結果:
- 統合前: 233行 + 282行 + 64行 = 579行（3ファイル）
- 統合後: 514行（1ファイル）
- ファイル削減率: 67%
- コード重複排除と保守性向上

Note: Important Tierファイルは500行制限のため、300行制限をバイパス
